Split monitor package into multiple files
Monitor, MonitorRunner and MonitorOutputProcessors are now defined in
separate files: monitor.py, runner.py and parser.py respectively
